play: revert the three ~static_cast<i32> casts - the bug they guarded is not real
cl 5.0 promotes every enumerator to int under `~`. Probed directly at /Od, /O1
and /O2 with named and anonymous enums over enumerator ranges 1, 0x80, 0x8000,
0xffff, 0x10000 and -1: every case gives the full 32-bit complement. All three
cited Play.cpp sites already emitted retail's exact bytes from the UNCAST
source (`83 60 40 fe` / `83 e1 fe`), and tree-wide there is no `andl $0xfffe`
in any obj - the only 0xfffe immediates anywhere are twelve cmpl/addl.

Reverting the casts is not neutral, it is a small WIN: 3470 -> 3471 exact.
They were dead weight against a cast metric ratcheted to 0.

The census that came out of the sweep is worth keeping: 602 `~` tokens
tree-wide, summing exactly - 345 destructors, 108 numeric literals, 96
~TILE_MASK_PX, 41 ~SPRITE_STATE_HIDDEN, 4 ~SPRITE_STATE_FLASHING, 3 already
cast, 2 ~variable, 1 ~WS_EX_TOPMOST, 1 ~member, 1 the operator~ in Enums.h.
Zero truncate. (The `&= ~[A-Z_]+;` regex I briefed finds only ~45 of the 149
named-constant sites - it misses every `x & ~TILE_MASK_PX` form.)

+++ src/Gruntz/Play.cpp
@@ -3749,7 +3749,7 @@ i32 CPlay::LoadCursorSprites(i32 frame, i32 flag) {
             return 0;
         }
         if (this->m_scrollSink != NULL) {
-            this->m_scrollSink->m_stateFlags &= ~static_cast<i32>(SPRITE_STATE_HIDDEN);
+            this->m_scrollSink->m_stateFlags &= ~SPRITE_STATE_HIDDEN;
         }
         this->m_cursorOffset.m_x = 0x10;
         this->m_cursorOffset.m_y = 0x10;
@@ -4179,7 +4179,7 @@ i32 CPlay::HandleDragMove(i32 a, i32 x, i32 y) {
                 }
             } else {
                 if (m_scrollSink != NULL) {
-                    m_scrollSink->m_stateFlags &= ~static_cast<i32>(SPRITE_STATE_HIDDEN);
+                    m_scrollSink->m_stateFlags &= ~SPRITE_STATE_HIDDEN;
                 }
             }
             CGameLevel* v = m_world->m_level;
@@ -4640,7 +4640,7 @@ i32 CPlay::ResetPlayState() {
         }
     }
     if (m_scrollSink != NULL) {
-        m_scrollSink->m_stateFlags &= ~static_cast<i32>(SPRITE_STATE_HIDDEN);
+        m_scrollSink->m_stateFlags &= ~SPRITE_STATE_HIDDEN;
     }
     m_inGame = 0;
     if (!PlaceStartGruntz()) {
